+++ tests/Fixture/Classes/WithMultipleParameters.php
@@ -1,0 +1,59 @@
+<?php
+
+namespace Rougin\Slytherin\Fixture\Classes;
+
+/**
+ * With Multiple Parameters
+ *
+ * @package Slytherin
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+class WithMultipleParameters
+{
+    /**
+     * @var array
+     */
+    protected $data = array();
+
+    /**
+     * @var array
+     */
+    protected $fields = array();
+
+    /**
+     * @var null
+     */
+    protected $lang = null;
+
+    /**
+     * @var null
+     */
+    protected $dir = null;
+
+    /**
+     * @param array       $data
+     * @param array       $fields
+     * @param string|null $lang
+     * @param string|null $dir
+     */
+    public function __construct($data = array(), $fields = array(), $lang = null, $dir = null)
+    {
+        $this->data = $data;
+
+        $this->fields = $fields;
+
+        $this->lang = $lang;
+
+        $this->dir = $dir;
+    }
+
+    /**
+     * Returns with a string of "With multiple parameters".
+     *
+     * @return string
+     */
+    public function index()
+    {
+        return 'With multiple parameters';
+    }
+}